* commit '2902421126db3683213ffabc521c2c2b7ac130b0':
  ALFSED-2061 фикс проблемы удаления вложения поручения/поручения с вложением. +фикс работы аплоудера на форме редактирования отчета соисполнителя.

+++ lecm-errands/repo/src/main/java/ru/it/lecm/errands/policy/ErrandsReportAttachmentAssociationPolicy.java
@@ -159,8 +159,8 @@ public class ErrandsReportAttachmentAssociationPolicy implements NodeServicePoli
         NodeRef executionTableData = nodeService.getTargetAssocs(errandDoc, ErrandsService.ASSOC_ERRANDS_TS_EXECUTION_REPORTS).get(0).getTargetRef();
         Boolean attachmentIsUsed = isAttachmentUsedInTable(coexecutorsTableData, ErrandsService.ASSOC_ERRANDS_TS_COEXECUTOR_ATTACHMENT, reportNodeRef, attachment) ||
                 isAttachmentUsedInTable(executionTableData, ErrandsService.ASSOC_ERRANDS_TS_EXECUTOR_ATTACHMENT, reportNodeRef, attachment);
-
-        if (!attachmentIsUsed) {
+        Boolean isPendingDeletion = nodeService.hasAspect(attachment, ContentModel.ASPECT_PENDING_DELETE);
+        if (!attachmentIsUsed && !isPendingDeletion) {
             documentAttachmentsService.deleteAttachment(attachment);
         }
     }
